refactor(profile): dedupe playlist suggestion handlers
