+++ apps/web/app/api/suggestions/playlist-fallback/[id]/approve/route.ts
@@ -1,74 +1,24 @@
 import { eq } from 'drizzle-orm';
 import { NextResponse } from 'next/server';
-import { z } from 'zod';
-import { getCachedAuth } from '@/lib/auth/cached';
 import { invalidateProfileCache } from '@/lib/cache/profile';
 import { db } from '@/lib/db';
-import { users } from '@/lib/db/schema/auth';
 import { creatorProfiles } from '@/lib/db/schema/profiles';
 import { captureError } from '@/lib/error-tracking';
-import { getFeaturedPlaylistFallbackCandidate } from '@/lib/profile/featured-playlist-fallback';
-
-const requestSchema = z.object({
-  profileId: z.string().uuid(),
-});
+import { getPlaylistFallbackRequestContext } from '../../_shared';
 
 export async function POST(
   request: Request,
   { params }: { params: Promise<{ id: string }> }
 ) {
   try {
-    const { id: playlistId } = await params;
-    const { userId } = await getCachedAuth();
+    const context = await getPlaylistFallbackRequestContext(request, params);
 
-    if (!userId) {
-      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
+    if (context instanceof NextResponse) {
+      return context;
     }
 
-    const body = await request.json();
-    const parsed = requestSchema.safeParse(body);
-
-    if (!parsed.success) {
-      return NextResponse.json(
-        {
-          success: false,
-          error: 'Invalid request',
-          details: parsed.error.flatten(),
-        },
-        { status: 400 }
-      );
-    }
-
-    const { profileId } = parsed.data;
-
-    const [profile] = await db
-      .select({
-        id: creatorProfiles.id,
-        clerkId: users.clerkId,
-        settings: creatorProfiles.settings,
-        usernameNormalized: creatorProfiles.usernameNormalized,
-      })
-      .from(creatorProfiles)
-      .innerJoin(users, eq(users.id, creatorProfiles.userId))
-      .where(eq(creatorProfiles.id, profileId))
-      .limit(1);
-
-    if (!profile || profile.clerkId !== userId) {
-      return NextResponse.json(
-        { success: false, error: 'Forbidden' },
-        { status: 403 }
-      );
-    }
-
-    const settings = (profile.settings ?? {}) as Record<string, unknown>;
-    const candidate = getFeaturedPlaylistFallbackCandidate(settings);
-
-    if (!candidate || candidate.playlistId !== playlistId) {
-      return NextResponse.json(
-        { success: false, error: 'Playlist suggestion not found' },
-        { status: 404 }
-      );
-    }
+    const { candidate, playlistId, profileId, settings, usernameNormalized } =
+      context;
 
     const nextSettings = {
       ...settings,
@@ -91,7 +41,7 @@ export async function POST(
       })
       .where(eq(creatorProfiles.id, profileId));
 
-    await invalidateProfileCache(profile.usernameNormalized);
+    await invalidateProfileCache(usernameNormalized);
 
     return NextResponse.json({
       success: true,
